requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
metaphor-python==0.1.16
langchain==0.0.274
python-dotenv
openai
